refactor: moved HttpRequestParams to its own files
