+++ src/agent/communicator/CMakeLists.txt
@@ -15,7 +15,7 @@ find_package(Boost REQUIRED COMPONENTS asio beast url)
 find_package(nlohmann_json CONFIG REQUIRED)
 find_path(JWT_CPP_INCLUDE_DIRS "jwt-cpp/base.h")
 
-add_library(Communicator src/communicator.cpp src/http_client.cpp)
+add_library(Communicator src/communicator.cpp src/http_client.cpp src/http_request_params.cpp)
 target_include_directories(Communicator PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/include SYSTEM INTERFACE ${JWT_CPP_INCLUDE_DIRS})
 target_link_libraries(Communicator PUBLIC Boost::asio PRIVATE Boost::beast Boost::url OpenSSL::SSL OpenSSL::Crypto Logger)
