+++ src/agent/communicator/include/ihttp_client.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <http_request_params.hpp>
 #include <ihttp_socket.hpp>
 
 #include <boost/asio.hpp>
@@ -13,45 +14,6 @@
 
 namespace http_client
 {
-    struct HttpRequestParams
-    {
-        boost::beast::http::verb Method;
-        std::string Host;
-        std::string Port;
-        std::string Endpoint;
-        bool Use_Https;
-        std::string Token;
-        std::string User_pass;
-        std::string Body;
-
-        HttpRequestParams(boost::beast::http::verb method,
-                          const std::string& host,
-                          const std::string& endpoint,
-                          const std::string& token = "",
-                          const std::string& user_pass = "",
-                          const std::string& body = "")
-            : Method(method)
-            , Host(host)
-            , Endpoint(endpoint)
-            , Token(token)
-            , User_pass(user_pass)
-            , Body(body)
-        {
-            boost::urls::url_view url(host);
-
-            Use_Https = url.scheme().empty() || url.scheme() == "https";
-            Host = url.host();
-            Port = url.port();
-        }
-
-        bool operator==(const HttpRequestParams& other) const
-        {
-            return Method == other.Method && Host == other.Host && Port == other.Port && Endpoint == other.Endpoint &&
-                   Token == other.Token && User_pass == other.User_pass && Body == other.Body &&
-                   Use_Https == other.Use_Https;
-        }
-    };
-
     class IHttpClient
     {
     public:
